pyproject.toml
[tool.poetry]
name = "cw6-drf"
version = "0.1.0"
description = ""
authors = ["Evgeny Egorov <eaegorov42@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
flake8 = "^7.1.1"
django-filter = "^24.3"
black = "^24.10.0"
django-crontab = "^0.7.1"
djangorestframework-simplejwt = "^5.3.1"
isort = "^5.13.2"
ipython = "^8.29.0"
coverage = "^7.6.4"
stripe = "^11.1.1"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.5.0"
requests = "^2.32.3"
forex-python = "^1.8"
redis = "^5.2.0"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
pytest = "^8.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
